remove log components, just format directly with fmtlib

+++ loader/src/ui/nodes/MDTextArea.cpp
@@ -320,7 +320,7 @@ struct MDParser {
 
             default:
                 {
-                    log::warn("Unhandled text type {}", type);
+                    log::warn("Unhandled text type {}", static_cast<int>(type));
                 }
                 break;
         }
@@ -413,7 +413,7 @@ struct MDParser {
 
             default:
                 {
-                    log::warn("Unhandled block enter type {}", type);
+                    log::warn("Unhandled block enter type {}", static_cast<int>(type));
                 }
                 break;
         }
@@ -516,7 +516,7 @@ struct MDParser {
 
             default:
                 {
-                    log::warn("Unhandled block leave type {}", type);
+                    log::warn("Unhandled block leave type {}", static_cast<int>(type));
                 }
                 break;
         }
@@ -573,7 +573,7 @@ struct MDParser {
 
             default:
                 {
-                    log::warn("Unhandled span enter type {}", type);
+                    log::warn("Unhandled span enter type {}", static_cast<int>(type));
                 }
                 break;
         }
@@ -627,7 +627,7 @@ struct MDParser {
 
             default:
                 {
-                    log::warn("Unhandled span leave type {}", type);
+                    log::warn("Unhandled span leave type {}", static_cast<int>(type));
                 }
                 break;
         }
